refactor: App.tsx 세부 컴포넌트 단위 분리
